mysite/requirements.txt
Django==5.1.2
psycopg==3.1.9
djangorestframework==3.15.2
django-cors-headers==4.7.0